front end web development lab 1
Create a web page that uses JavaScript to interact with the DOM in the following ways:

manipulate at least one CSS property through the DOM
manipulate at least one DOM property that is NOT a CSS property
responds to at least two different types of events using event listeners(using addEventListener()).
Your JavaScript should be in a function in an external js file.